apps/api/pyproject.toml
[project]
name = "polaris-api"
version = "0.0.0"
description = "FastAPI control plane for Polaris."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.0.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "alembic>=1.13.0",
  "asyncpg>=0.29.0",
  "psycopg[binary]>=3.2.0",
  "redis>=5.0.0",
  "httpx>=0.27.0",
  "PyJWT>=2.9.0",
  "PyYAML>=6.0",
  "aioboto3>=13.0.0",
  "fastmcp>=3.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "respx>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/polaris_api"]

[tool.ruff]
line-length = 100
target-version = "py312"
